Move helper function to Utils::Fs
